added: window render thread

+++ src/Core/Rendering.hpp
@@ -1,8 +1,7 @@
 #pragma once
 
 #include <Core/Types.hpp>
-#include <Core/Rendering.hpp>
 
-constexpr f32 FpsToHz(int fps) {
-    return float(1000.0) / float(fps);
+constexpr u32 FpsToHz(int fps) {
+    return u32(1000.0 / float(fps));
 }
